rename(api): chat -> chatRoom
