+++ src/app/(route)/chat/_components/ChatItem/ChatItem.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import { ChatRoom } from "@/api/fetch/chat/types/ChatListType";
+import { ChatRoom } from "@/api/fetch/chatRoom/types/ChatListType";
 import { Icon } from "@/components";
 import { formatDate } from "@/utils";
 import { MOCK_IMAGES } from "./MOCK_IMAGES";
